move common to actions
